+++ src/RenderingDriver/RenderingDriverGL/ContextGL.Buffer.cpp
@@ -6,12 +6,13 @@
 namespace Xe { namespace Graphics {
 	bool CContextGLCommon::CreateBuffer(IBuffer **ppBuffer, const BufferDesc& desc, DataDesc* pData)
 	{
+		ASSERT(!"CreateBuffer not implemented yet.");
 		return false;
 	}
 
 	void CContextGLCommon::SelectBuffer(IBuffer *pBuffer)
 	{
-
+		ASSERT(!"SelectBuffer not implemented yet.");
 	}
 } }